Move config_loader into mod
