--- src/tracer/src/config/config_loader.rs
@@ -1,32 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use crate::cloud_providers::aws::config::AwsConfig;
-use crate::cloud_providers::aws::types::aws_region::AwsRegion;
-use crate::common::target_process::Target;
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct Config {
-    pub api_key: String,
-    pub process_polling_interval_ms: u64,
-    pub batch_submission_interval_ms: u64,
-    pub process_metrics_send_interval_ms: u64,
-    pub file_size_not_changing_period_ms: u64,
-    pub new_run_pause_ms: u64,
-    pub targets: Vec<Target>,
-
-    pub aws_init_type: AwsConfig,
-    pub aws_region: AwsRegion,
-
-    pub database_secrets_arn: Option<String>,
-    pub database_host: Option<String>,
-    pub database_name: String,
-
-    pub grafana_workspace_url: String,
-    pub server: String,
-
-    pub config_sources: Vec<String>,
-    pub sentry_dsn: Option<String>,
-
-    pub log_forward_endpoint_dev: Option<String>,
-    pub log_forward_endpoint_prod: Option<String>,
-}
